исправлены баги

+++ internal/service/place/service.go
@@ -8,12 +8,12 @@ import (
 
 type PlaceService interface {
 	CreatePlace(ctx context.Context, place placeDomain.Place) error
-	GetPlaces(context.Context, float64, float64, float64, float64) ([]placeDomain.Place, error)
+	GetPlaces(context.Context, float32, float32, float32, float32) ([]placeDomain.Place, error)
 }
 
 type PlaceStorage interface {
 	CreatePlace(ctx context.Context, place placeDomain.Place) error
-	GetPlaces(context.Context, float64, float64, float64, float64) ([]placeDomain.Place, error)
+	GetPlaces(context.Context, float32, float32, float32, float32) ([]placeDomain.Place, error)
 }
 
 type service struct {
@@ -32,10 +32,10 @@ func (s *service) CreatePlace(ctx context.Context, place placeDomain.Place) erro
 
 func (s *service) GetPlaces(
 	ctx context.Context,
-	minLongitude float64,
-	maxLongitude float64,
-	minLatitude float64,
-	maxLatitude float64,
+	minLongitude float32,
+	maxLongitude float32,
+	minLatitude float32,
+	maxLatitude float32,
 ) ([]placeDomain.Place, error) {
 	return s.placeStorage.GetPlaces(ctx, minLongitude, maxLongitude, minLatitude, maxLatitude)
 }
